requirements.txt
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
attr==0.3.1
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.20.0
bleach==5.0.1
boto==2.49.0
boto3==1.34.109
botocore==1.34.109
boxing==0.1.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.7.1
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
exceptiongroup==1.2.1
expiringdict==1.2.2
filelock==3.14.0
fsspec==2024.5.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-logging==3.10.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
humansignal-drf-yasg==1.21.9
idna==3.7
ijson==3.2.3
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==3.2.0
label-studio==1.12.0.post0
label-studio-converter==0.0.58
label-studio-tools==0.0.4
launchdarkly-server-sdk==8.2.1
llvmlite==0.42.0
lockfile==0.12.2
lxml==5.2.2
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
nltk==3.6.7
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.30.1
openai-whisper==20231117
ordered-set==4.0.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.15
pydub==0.25.1
pyRFC3339==1.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.1
redis==3.5.3
regex==2024.5.15
requests==2.31.0
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.1
safetensors==0.4.3
semver==3.0.2
sentry-sdk==2.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.18.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
webencodings==0.5.1
xmljson==0.2.0
